+++ game/content/ghplots/tarot_sockets.py
@@ -1,3 +1,4 @@
+import pbge.memos
 from pbge.plots import Plot, PlotState
 from game.content import ghwaypoints,ghterrain,plotutility,backstory
 import gears
@@ -17,9 +18,7 @@ from . import dd_combatmission
 import collections
 from . import missionbuilder
 
-from game import memobrowser
-
-Memo = memobrowser.Memo
+Memo = pbge.memos.Memo
 
 #   **************************
 #   ***  MT_SOCKET_Accuse  ***