+++ pbge/plots.py
@@ -1,6 +1,6 @@
 import random
 from . import dialogue
-
+from .memos import Memo
 
 class PlotError( Exception ):
     """Plot init will call this if initialization impossible."""
@@ -48,25 +48,6 @@ class PlotState( object ):
         # Why return self? Because this function will often be called straight
         # from the generator.
         return self
-
-
-class Memo(object):
-    def __init__(self, text, location=None):
-        self._text = text
-        self._location = location
-
-    def __str__(self):
-        if not self._location:
-            return self._text
-        if hasattr(self._location, 'get_root_scene'):
-            root = self._location.get_root_scene()
-            if root is not self._location:
-                loc = "{} at {}".format(self._location, root)
-            else:
-                loc = str(self._location)
-        else:
-            loc = str(self._location)
-        return "{}\n\nLocation: {}".format(self._text, loc)
 
 
 class RumorMemoEffect(object):